Add files via upload
The lexical decision task includes a list of 46 items, of which 23 are words and 23 are non-words. 
The words are shown on the screen for 0.5 seconds and are separated by the presentation of a fixation cross, which is displayed for 0.725 seconds. 
At the end of this task, which was originally designed to be a practice trail of an ongoing task of a prospective memory task, participants will see a screen with their score and will be asked if they want to do the task again or not.